feat(83): refactor browser fixture and remove unused configurations
